till task 3 features upperlowerquarterhistograms
TODO : distance metric for task 3.

if intersec > 0.01 do intersec * 100 else intersec  = 1.

+++ app/include/distanceutils.h
@@ -51,7 +51,7 @@ class DistanceFinder{
         std::string targetPath;
         std::string distanceName;
         distanceMethod distanceComputer;
-        std::vector<float> distances;
+        std::vector<double> distances;
         std::map<std::string, std::vector<std::vector<double>>> featureMap;
         std::vector<double> distancesSorted;
         std::vector<std::string> imPathsDistSorted;
